unsetting PWD will no longer segfault and the prompt is redisplayed after SIGINT
