+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: aprevrha <aprevrha@student.42vienna.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/15 00:10:11 by aprevrha          #+#    #+#             */
-/*   Updated: 2024/05/21 15:16:47 by aprevrha         ###   ########.fr       */
+/*   Updated: 2024/05/22 00:48:55 by aprevrha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -195,6 +195,8 @@ char				*read_next_line(int fd);
 char				*vstrjoin(int nbr_of_strs, ...);
 void				set_is_pipeline(t_list *cmds);
 bool				dont_run(const t_cmd *cmd);
+
+char				*get_prompt(t_evars *evars);
 
 // loop.c
 char				*ft_readline(char *prompt, int tty);
